backend/pyproject.toml
[project]
name = "agent-communications-dashboard"
version = "0.1.0"
description = "Backend service for agent communications dashboard"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "opentelemetry-exporter-otlp-proto-http>=1.36.0",
    "opentelemetry-proto>=1.36.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_monitor"]

[dependency-groups]
dev = [
    "poethepoet>=0.31.1",
    "mypy>=1.9.0",
    "ruff>=0.11.2",
    "pytest>=8.1.1",
    "bandit[toml]>=1.7.8",
    "import-linter>=2.0"
]

[tool.poe.tasks]
test = "pytest --strict-config"
test-watch = "ptw ."

format = "ruff format"
lint = "ruff check --fix"
ruff = ["format", "lint"]

bandit = "bandit -c pyproject.toml -r ."
mypy = "mypy ."
lint-imports = "lint-imports"

check = ["mypy", "bandit", "ruff", "lint-imports", "test"]

[tool.importlinter]
root_packages = [
    "agent_monitor",
]

[tool.ruff]
# Allow lines to be as long as 120.
line-length = 120


[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/
# I001 --> enable import sorting rules
# Q --> enable flake8-quotes rules for consistent quote style
# N --> pep8-naming for proper naming conventions
# PT --> flake8-pytest-style for pytest best practices
# TID251 --> banned relative imports
# DTZ --> datetime timezone awareness
extend-select = ["Q", "N", "I001", "PT", "TID251", "DTZ"]

[tool.bandit]
exclude_dirs = [".cache", ".venv"]
skips = [
    "B104", # hardcoded_bind_all_interfaces: Accept this for now.
]

[tool.bandit.assert_used]
skips = ["*_test.py", "*test_*.py"]

